Test: Fix integer type issues for updated Catch.

+++ unit_tests/test_IntrusiveHashMap.cc
@@ -99,7 +99,7 @@ TEST_CASE("IntrusiveHashMap", "[libts][IntrusiveHashMap]")
 
   size_t nb = map.bucket_count();
   std::bitset<64> marks;
-  for (int i = 1; i <= 63; ++i) {
+  for (size_t i = 1; i <= 63; ++i) {
     std::string name;
     swoc::bwprint(name, "{} squared is {}", i, i * i);
     Thing *thing = new Thing(name);
